requirements.txt
alembic==0.9.1
appdirs==1.4.3
beautifulsoup4==4.5.3
click==6.7
defusedxml==0.5.0
django-htmlmin==0.10.0
Flask==0.12.1
Flask-Assets==0.12
Flask-Bower==1.2.1
Flask-Login==0.4.0
flask-marshmallow==0.7.0
Flask-Migrate==2.0.3
Flask-Script==2.0.5
Flask-SQLAlchemy==2.2
html5lib==0.999999999
itsdangerous==0.24
Jinja2==2.9.6
Mako==1.0.6
MarkupSafe==1.0
marshmallow==2.13.4
marshmallow-sqlalchemy==0.13.1
mccabe==0.6.1
oauthlib==2.0.2
packaging==16.8
pycodestyle==2.3.1
pyflakes==1.5.0
PyJWT==1.4.2
pyparsing==2.2.0
python-editor==1.0.3
python3-openid==3.1.0
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
social-auth-app-flask==1.0.0
social-auth-app-flask-sqlalchemy==1.0.1
social-auth-core==1.2.0
social-auth-storage-sqlalchemy==1.0.1
SQLAlchemy==1.1.9
webassets==0.12.1
webencodings==0.5.1
Werkzeug==0.12.1
